feat: Define a State trait
A State trait has 3 methods; enter, process, exit. And these methods
can be specialized by the implementations. There is a notion of a
state machine which has the notion of the current state. Here there is
a structure with cur_state for which you can dynamically invoke the
State trait methods.